used Solid principal of OOP
