+++ main.rb
@@ -1,4 +1,5 @@
 require_relative './app'
+require_relative './modules/choose_operation'
 
 @app = App.new
 
@@ -27,23 +28,6 @@ def start_program
         choose_operation(choice)
       end
     end
-  end
-end
-
-def choose_operation(opt)
-  case opt
-  when 1
-    @app.list_books
-  when 2
-    @app.list_peoples
-  when 3
-    @app.create_person
-  when 4
-    @app.create_book
-  when 5
-    @app.create_rental
-  when 6
-    @app.list_rentals
   end
 end
